Break the content of index.tsx into 2 separate files
